Move CB to different file
